Extract format_codes to its own file
